move the get post details call to its own component.
