+++ src/components/mod.rs
@@ -1,2 +1,3 @@
 pub(crate) mod create_post;
 pub(crate) mod footer;
+pub(crate) mod post_details;
